Change telegram_id to BigInteger for larger values

+++ app/models/user.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer, String
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import relationship
+from sqlalchemy import BigInteger, Boolean, Column, DateTime, ForeignKey, Integer, String
+from sqlalchemy.orm import declarative_base, relationship
 from sqlalchemy.sql import func
 
 Base = declarative_base()
@@ -12,7 +11,8 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    # Telegram IDs can exceed int32 range
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     first_name = Column(String(128), nullable=True)
     last_name = Column(String(128), nullable=True)
     username = Column(String(64), nullable=True)
